[SECURITY] Avoid download from fallback storage in FileDownloadController
The `FileDownloadController` allows selecting several files
from a given folder for download as zip file. Files for
download are passed with the schema
`<storage-id>:/folder/filename.extension` as POST data to
the controller. The controller retrieves the file from FAL
and provides it as download when the backend user has
access to it. Files or folders in the FAL fallback storage
`0` will currently be resolved, validated for storage
access permissions with `checkFileActionPermission()` and
provided for download if access is granted. This could
lead to download of files with sensitive data like
logfiles.

With this change, downloads of files from the fallback
storage are denied in general for all users. In addition,
functional tests have been added to cover all access
check scenarios of the controller.

Resolves: #108749
Releases: main, 14.3, 13.4
Security-Bulletin: TYPO3-CORE-SA-2026-013
Security-References: CVE-2026-49742
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/94416
Tested-by: Oliver Hader <oliver.hader@typo3.org>

+++ typo3/sysext/filelist/Classes/Controller/FileDownloadController.php
@@ -101,7 +101,9 @@ class FileDownloadController
         }
         $zipFile->close();
         $response = $this->createResponse($zipFileName, $filesAdded);
-        unlink($zipFileName);
+        if ($filesAdded > 0) {
+            unlink($zipFileName);
+        }
         return $response;
     }
 
@@ -132,6 +134,10 @@ class FileDownloadController
         foreach ($items as $itemIdentifier) {
             $fileOrFolderObject = $this->resourceFactory->retrieveFileOrFolderObject($itemIdentifier);
             if ($fileOrFolderObject === null) {
+                continue;
+            }
+            // Files from fallback storage must be skipped in general
+            if ($fileOrFolderObject->getStorage()->isFallbackStorage()) {
                 continue;
             }
             $baseIdentifier = dirname($fileOrFolderObject->getIdentifier());
